requirements.txt
pandas>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
twilio>=8.2.0
python-dotenv>=0.19.0
google-cloud-logging>=3.5.0
google-cloud-storage>=2.10.0 